refactor: 재사용을 위한 사용자 json 파일의 이미지 경로 수정
